Simplify the SessionTracker to the essentials
The SessionTracker is now only used to send sessions within the same
request. There is no provision for storing data across requests
