+++ tests/SessionTrackerTest.php
@@ -1,22 +1,30 @@
 <?php
 
-namespace Bugsnag\Tests;
+namespace Bugsnag\Tests\SessionTracker;
 
 use Bugsnag\Configuration;
 use Bugsnag\HttpClient;
-use Bugsnag\SessionTracker;
+use Bugsnag\SessionTracker\SessionTracker;
+use Bugsnag\SessionTracker\SessionTrackerInterface;
+use Bugsnag\Tests\TestCase;
 use InvalidArgumentException;
 use PHPUnit\Framework\MockObject\MockObject;
-use RuntimeException;
-use stdClass;
 
 class SessionTrackerTest extends TestCase
 {
-    /** @var SessionTracker */
+    /**
+     * @var SessionTracker
+     */
     private $sessionTracker;
-    /** @var Configuration&MockObject */
+
+    /**
+     * @var Configuration&MockObject
+     */
     private $config;
-    /** @var HttpClient&MockObject */
+
+    /**
+     * @var HttpClient&MockObject
+     */
     private $client;
 
     public function setUp()
@@ -26,6 +34,7 @@ class SessionTrackerTest extends TestCase
             ->setConstructorArgs(['example-api-key'])
             ->getMock();
 
+        /** @var HttpClient&MockObject */
         $this->client = $this->getMockBuilder(HttpClient::class)
             ->disableOriginalConstructor()
             ->getMock();
@@ -33,131 +42,32 @@ class SessionTrackerTest extends TestCase
         $this->sessionTracker = new SessionTracker($this->config, $this->client);
     }
 
-    public function testSendSessionsEmpty()
+    public function testItImplementsTheSessionTrackerInterface()
+    {
+        $this->assertInstanceOf(
+            SessionTrackerInterface::class,
+            $this->sessionTracker
+        );
+    }
+
+    public function testSendSessionsDoesNothingIfThereIsNoDataToSend()
     {
         $this->client->expects($this->never())->method('sendSessions');
 
         $this->sessionTracker->sendSessions();
     }
 
-    public function testSendSessionsShouldNotNotify()
+    public function testSendSessionsDoesNothingIfReleaseStageIsIgnored()
     {
-        $numberOfCalls = 0;
-
-        $this->sessionTracker->setStorageFunction(function ($key, $value = null) use (&$numberOfCalls) {
-            $numberOfCalls++;
-
-            if ($value === null) {
-                $this->assertSame(1, $numberOfCalls, 'Expected the first call to be a read ($value === null)');
-
-                return ['2000-01-01T00:00:00' => 1];
-            }
-
-            $this->assertSame(2, $numberOfCalls, 'Expected the second call to be a write ($value === [])');
-            $this->assertSame([], $value, 'Expected the second call to be a write ($value === [])');
-        });
-
         $this->config->expects($this->once())->method('shouldNotify')->willReturn(false);
         $this->client->expects($this->never())->method('sendSessions');
 
-        $this->sessionTracker->sendSessions();
-
-        $this->assertSame(2, $numberOfCalls, 'Expected there to be two calls to the session storage function');
-    }
-
-    /**
-     * @param mixed $returnValue
-     *
-     * @return void
-     *
-     * @dataProvider storageFunctionEmptyReturnValueProvider
-     */
-    public function testSendSessionsReturnsEarlyWhenGetSessionCountsReturnsAValueThatsNotAPopulatedArray($returnValue)
-    {
-        $this->sessionTracker->setStorageFunction(function () use ($returnValue) {
-            return $returnValue;
-        });
-
-        $this->config->expects($this->never())->method('shouldNotify');
-        $this->config->expects($this->never())->method('getNotifier');
-        $this->config->expects($this->never())->method('getDeviceData');
-        $this->config->expects($this->never())->method('getAppData');
-        $this->config->expects($this->never())->method('getApiKey');
-        $this->client->expects($this->never())->method('sendSessions');
-
+        $this->sessionTracker->startSession();
         $this->sessionTracker->sendSessions();
     }
 
-    /**
-     * @param mixed $returnValue
-     *
-     * @return void
-     *
-     * @dataProvider storageFunctionEmptyReturnValueProvider
-     */
-    public function testStartSessionDoesNotDeliverSessionsWhenLastSentIsNotAnInteger($returnValue)
+    public function testSessionsCanBeSentExplicitly()
     {
-        $this->sessionTracker->setStorageFunction(function ($key) use ($returnValue) {
-            // We only care about the "last sent" value here
-            if ($key === 'bugsnag-sessions-last-sent') {
-                return $returnValue;
-            }
-
-            return null;
-        });
-
-        $this->config->expects($this->never())->method('shouldNotify');
-        $this->config->expects($this->never())->method('getNotifier');
-        $this->config->expects($this->never())->method('getDeviceData');
-        $this->config->expects($this->never())->method('getAppData');
-        $this->config->expects($this->never())->method('getApiKey');
-        $this->client->expects($this->never())->method('sendSessions');
-
-        $this->sessionTracker->startSession();
-    }
-
-    /**
-     * @param mixed $returnValue
-     *
-     * @return void
-     *
-     * @dataProvider storageFunctionEmptyReturnValueProvider
-     */
-    public function testStartSessionDoesNotDeliverSessionsWhenGetSessionCountsReturnsAValueThatsNotAPopulatedArray($returnValue)
-    {
-        $this->sessionTracker->setStorageFunction(function ($key) use ($returnValue) {
-            return $returnValue;
-        });
-
-        $this->config->expects($this->never())->method('shouldNotify');
-        $this->config->expects($this->never())->method('getNotifier');
-        $this->config->expects($this->never())->method('getDeviceData');
-        $this->config->expects($this->never())->method('getAppData');
-        $this->config->expects($this->never())->method('getApiKey');
-        $this->client->expects($this->never())->method('sendSessions');
-
-        $this->sessionTracker->startSession();
-    }
-
-    public function storageFunctionEmptyReturnValueProvider()
-    {
-        return [
-            'null' => [null],
-            'empty array' => [[]],
-            'int' => [1],
-            'float' => [1.2],
-            'bool (true)' => [true],
-            'bool (false)' => [false],
-            'object' => [new stdClass()],
-        ];
-    }
-
-    public function testSendSessionsSuccess()
-    {
-        $this->sessionTracker->setStorageFunction(function () {
-            return ['2000-01-01T00:00:00' => 1];
-        });
-
         $this->config->expects($this->once())->method('shouldNotify')->willReturn(true);
         $this->config->expects($this->once())->method('getNotifier')->willReturn('test_notifier');
         $this->config->expects($this->once())->method('getDeviceData')->willReturn('device_data');
@@ -169,7 +79,10 @@ class SessionTrackerTest extends TestCase
                 && $payload['device'] === 'device_data'
                 && $payload['app'] === 'app_data'
                 && count($payload['sessionCounts']) === 1
-                && $payload['sessionCounts'][0]['startedAt'] === '2000-01-01T00:00:00'
+                && preg_match(
+                    '/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$/',
+                    $payload['sessionCounts'][0]['startedAt']
+                )
                 && $payload['sessionCounts'][0]['sessionsStarted'] === 1;
         };
 
@@ -177,25 +90,22 @@ class SessionTrackerTest extends TestCase
             ->method('sendSessions')
             ->with($this->callback($expectCallback));
 
+        // Set the lastSent property of the SessionTracker so that 'startSession'
+        // doesn't immediately send sessions
+        $setLastSent = function () {
+            $this->lastSent = time();
+        };
+
+        $setLastSent = $setLastSent->bindTo($this->sessionTracker, $this->sessionTracker);
+
+        $setLastSent();
+
+        $this->sessionTracker->startSession();
         $this->sessionTracker->sendSessions();
     }
 
-    public function testStartSessionsSuccess()
+    public function testSessionsAreSentOnStartSessionIfNotRecentlySent()
     {
-        $session = [];
-
-        $this->sessionTracker->setStorageFunction(function ($key, $value = null) use (&$session) {
-            if (!isset($session[$key])) {
-                $session[$key] = null;
-            }
-
-            if ($value === null) {
-                return $session[$key];
-            }
-
-            $session[$key] = $value;
-        });
-
         $this->config->expects($this->once())->method('shouldNotify')->willReturn(true);
         $this->config->expects($this->once())->method('getNotifier')->willReturn('test_notifier');
         $this->config->expects($this->once())->method('getDeviceData')->willReturn('device_data');
@@ -221,122 +131,10 @@ class SessionTrackerTest extends TestCase
         $this->sessionTracker->startSession();
     }
 
-    public function testSetLockFunctionsThrowsWhenBothFunctionsAreNotCallable()
-    {
-        $this->expectedException(InvalidArgumentException::class, 'Both lock and unlock functions must be callable');
-
-        $this->sessionTracker->setLockFunctions(null, function () {});
-    }
-
-    public function testSetLockFunctionsThrowsWhenLockIsNotCallable()
-    {
-        $this->expectedException(InvalidArgumentException::class, 'Both lock and unlock functions must be callable');
-
-        $this->sessionTracker->setLockFunctions(null, function () {});
-    }
-
-    public function testSetLockFunctionsThrowsWhenUnlockIsNotCallable()
-    {
-        $this->expectedException(InvalidArgumentException::class, 'Both lock and unlock functions must be callable');
-
-        $this->sessionTracker->setLockFunctions(function () {}, null);
-    }
-
-    public function testSetLockFunctionsSucceedsWhenBothFunctionsAreCallable()
-    {
-        $locked = false;
-        $lockWasCalled = false;
-        $unlockWasCalled = false;
-
-        $this->sessionTracker->setLockFunctions(
-            function () use (&$locked, &$lockWasCalled) {
-                $locked = true;
-                $lockWasCalled = true;
-            },
-            function () use (&$locked, &$unlockWasCalled) {
-                $locked = false;
-                $unlockWasCalled = true;
-            }
-        );
-
-        $session = [];
-
-        $this->sessionTracker->setStorageFunction(function ($key, $value = null) use (&$session) {
-            if (!isset($session[$key])) {
-                $session[$key] = null;
-            }
-
-            if ($value === null) {
-                return $session[$key];
-            }
-
-            $session[$key] = $value;
-        });
-
-        $this->config->expects($this->once())->method('shouldNotify')->willReturn(true);
-
-        $this->sessionTracker->startSession();
-
-        $this->assertFalse($locked, 'Expected not to be locked after sending sessions');
-        $this->assertTrue($lockWasCalled, 'Expected the `lockFunction` to be called');
-        $this->assertTrue($unlockWasCalled, 'Expected the `unlockFunction` to be called');
-    }
-
-    public function testSessionShouldBeUnlockedAfterAnException()
-    {
-        $locked = false;
-        $lockWasCalled = false;
-        $unlockWasCalled = false;
-
-        $this->sessionTracker->setLockFunctions(
-            function () use (&$locked, &$lockWasCalled) {
-                $locked = true;
-                $lockWasCalled = true;
-            },
-            function () use (&$locked, &$unlockWasCalled) {
-                $locked = false;
-                $unlockWasCalled = true;
-            }
-        );
-
-        $this->sessionTracker->setStorageFunction(function () {
-            throw new RuntimeException('Something went wrong!');
-        });
-
-        $this->config->expects($this->never())->method('shouldNotify');
-
-        $e = null;
-
-        try {
-            $this->sessionTracker->startSession();
-        } catch (RuntimeException $e) {
-            $this->assertSame('Something went wrong!', $e->getMessage());
-        }
-
-        $this->assertNotNull($e, 'Expected a RuntimeException to be thrown');
-        $this->assertFalse($locked, 'Expected not to be locked after failing to send sessions');
-        $this->assertTrue($lockWasCalled, 'Expected the `lockFunction` to be called');
-        $this->assertTrue($unlockWasCalled, 'Expected the `unlockFunction` to be called');
-    }
-
     public function testSetRetryFunctionThrowsWhenNotGivenACallable()
     {
         $this->expectedException(InvalidArgumentException::class, 'The retry function must be callable');
 
         $this->sessionTracker->setRetryFunction(null);
-    }
-
-    public function testSetStorageFunctionThrowsWhenNotGivenACallable()
-    {
-        $this->expectedException(InvalidArgumentException::class, 'Storage function must be callable');
-
-        $this->sessionTracker->setStorageFunction(null);
-    }
-
-    public function testSetSessionFunctionThrowsWhenNotGivenACallable()
-    {
-        $this->expectedException(InvalidArgumentException::class, 'Session function must be callable');
-
-        $this->sessionTracker->setSessionFunction(null);
     }
 }
